Admin beta updates
Completed userSettings page - Changes to the userSettings file are now updated, and the previous userSettings file is archived.
Added default password forced change on initial login to Admin Dashboard
The site visitors' Map is fully functional. This is totally anonymous and is considered as such by the GDPR as it is only one piece of information and is not enough in the "PII puzzle" to be attached to a person as we are only using the IP address to the City/Region/Country level, then discarding the IP address.
Added the capability for the admin to keep his local IP from being counted as "visits" to his website by adding two new fields to the userSettings file, stripIP and localIP
Removed excess comments from the userSettings file
Admin Dashboard now displays the current version of the DivumWX software being run in prep for the update checker.