0.8.21
dvmAdvisoryOutlookTop.php.
Redesigned the units switching for the outlook summary which displays when an advisory is not in force. My original design switched on the location of the weather station which was a false premiss. It now switches depending on the units selected by the viewer - ie a viewer in USA viewing a station in Europe sees the outlook temperature in Farenheit if that is the viewer's choice. Previously a USA viewer would see the temperature in Celsius even if US customary units were their preference.