Database correction for DvM Dashboard
Updated baseline db
Modified index.php to verify that the database has the required new tables on attempted login
Preliminary work on the missing "total" variable in dvmDashboard.php, which is causing errors in the browser console. Note: This error is not a breaking error; it is just annoying at this point.