Major update to single CSS file
Significantly there is only 1 CSS file to use.

divumwx.main.css

The old stuff is at the top of the file below the variables. The start and finish are commented, likewise the new section which is at the bottom.

In order to deplete and clean the old stuff the rules are: -

1. If you are making amendments to old CSS, when the amendment is complete move it to the bottom of the new section but above all the @media queries and comment that it has been moved and what it was for.

2. If you are creating new CSS, again place it in the bottom above @media section and comment that it is new code and what it is for.

3. If you are amending CSS in the new section, comment when it was done.

4. I will regularly tidy up and move things around and hopefully will end up with a much more compact and smaller file which is much easier to wade around in.